requirements.txt
aiohttp==3.8.3
aiokafka==0.8.0
asyncpg==0.27.0
alembic==1.9.1
python-decouple==3.6
